requirements.txt
numpy==2.3.5
scipy==1.17.1
pandas==3.0.1
scikit-learn==1.8.0
PyYAML==6.0.3
tqdm==4.67.3
requests==2.32.5
psutil==7.2.2
nltk==3.9.3
rouge_score==0.1.2
sentencepiece==0.2.1
safetensors==0.7.0
huggingface_hub==1.7.2
transformers==5.3.0
accelerate==1.13.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
matplotlib

